refactor(kilo-chat): split message bubble content
